+++ CHANGELOG.md
@@ -15,9 +15,11 @@ and this project adheres to
 - 📝(project) add troubleshoot doc #1066
 - 📝(project) add system-requirement doc #1066
 - 🔧(front) configure x-frame-options to DENY in nginx conf #1084
--  (doc) add documentation to install with compose #855
 - ✨(backend) allow to disable checking unsafe mimetype on attachment upload
 - ✨Ask for access #1081
+- ✨(doc) add documentation to install with compose #855
+- ✨ Give priority to users connected to collaboration server 
+  (aka no websocket feature) #1093 
 
 ### Changed